moved site, sites, and vs30 to util from calc
