+++ src/org/opensha2/util/Site.java
@@ -1,4 +1,4 @@
-package org.opensha2.calc;
+package org.opensha2.util;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
@@ -6,11 +6,10 @@ import static com.google.common.base.Preconditions.checkState;
 import static org.opensha2.data.Data.checkInRange;
 import static org.opensha2.internal.GeoJson.validateProperty;
 
+import org.opensha2.calc.CalcConfig;
 import org.opensha2.geo.Location;
 import org.opensha2.gmm.GroundMotionModel;
 import org.opensha2.internal.GeoJson;
-import org.opensha2.util.Named;
-import org.opensha2.util.NamedLocation;
 
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableSet;
